* nasl_text_utils.c (nasl_int): Define r variable as long int instead of int to support bigger numbers.
git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-scanner@28093 423fd1db-d629-0410-8442-d21db03e70f4

+++ nasl/nasl_text_utils.c
@@ -1355,7 +1355,7 @@ nasl_str_replace (lex_ctxt * lexic)
 tree_cell *
 nasl_int (lex_ctxt * lexic)
 {
-  int r = get_int_var_by_num (lexic, 0, 0);
+  long int r = get_int_var_by_num (lexic, 0, 0);
   tree_cell *retc;
 
   retc = alloc_tree_cell (0, NULL);
